fix(migration): ScheduledTasks.Interval column type AsInt32 (IN-02)
The Interval column was declared `AsDouble()` while the backing model
(`Jobs/ScheduledTask.Interval`) is `int`. Dapper round-trips the value
fine on SQLite (loose typing), but the schema declared more precision
than the domain model would ever use, and on Postgres this materialized
as `double precision` for an integer minute count.

No other migration alters this column, no caller assumes double
precision, and Postgres has no data dependency on the wider type — so
per pre-v1 dev-migration policy this is an edit-001-in-place change.
Existing dev databases require a fresh DB wipe (standing approval per
project memory) for the schema change to land.

Refs #147 (IN-02).

+++ src/NzbDrone.Core/Datastore/Migration/001_mangarr_baseline.cs
@@ -58,7 +58,7 @@ namespace NzbDrone.Core.Datastore.Migration
 
             Create.TableForModel("ScheduledTasks")
                 .WithColumn("TypeName").AsString().Unique()
-                .WithColumn("Interval").AsDouble()
+                .WithColumn("Interval").AsInt32()
                 .WithColumn("LastExecution").AsDateTime()
                 .WithColumn("LastStartTime").AsDateTime().Nullable();
